Client: fix single-server case
A "for BACKUPSERVER in $backupservers[@]" style loop was used in both
zfsbackup-client as well as zfsbackup-create-source. If there was only a
single, unnamed backupserver (which would be the case in legacy
installations created before multi-server support was introduced), the
backupservers array only contained a single empty string as an element.
$backupservers[@] thus expanded to the empty string, causing the for loop
not to run (and no backups to be performed).

The "$backupservers[@]" expansion has to take place within quotes for the
for loop to be executed once (with BACKUPSERVER="").