zfsbackup-client: remove support for snapuuid -- zfs snapshots don't seem to have one, and if we're backing up ephemeral snapshots, the uuid isn't terribly useful anyway.
zfsbackup-client: improve metadata logging; for example, try to generate a
meaningful 'fsuuid' for the FS being backed up if one isn't provided.

zfsbackup-client: export zbSOURCENAME, rsync_options and starttime variables
for pre-client and post-client scripts.

zfsbackup-client: separately log (into files in the source.d directory) the
exit status of various subprocesses such as check, pre-client, pre-client.d,
rsync, post-client, post-client.d.

client/mksource.d/zfs-set-korn-zfsbackup-config: add comment debating what
to do if --no-xdev was specified.

README.md: minor updates

make-snapshot: instead of usedbydataset, use lrefer when checking if minsize
has been reached.  This should be closer to the intuitive meaning of 'how
much data is in the snapshot' when compression is enabled or when
copies > 1.

make-snapshot: try to invoke syncfs() on the backup target fs in the hope
that 'usedbydataset' and the used inode count reported by df -i will be
correct when it completes.

zfsbackup-create-source: support 'timelimit' in client-defaults (hardlink it
to newly created source config).

zfsbackup-client: add support for the `timelimit` option, which forces rsync
to be killed after a set amount of time, even if it is still transferring.

client/zfsbackup-sv: wrap zfsbackup-client with loadwatch(1) if it's
available.