Allow arguments to `zfs snapshot` to be passed through `make-snapshot`
While it's not the default and not supported by the
`zfsbackup-create-source` script, it might make sense to manually create a
hierarchy of zfs filesystems to backup a single host to, e.g.

```
/backup
|/host
||/var
||/home
||/srv
|||/postgres
|||/www
```

could all be separate filesystems (perhaps with different dedup and
compression settings), with only a single zfsbackup source existing on the
host for the entire tree that backs up everything, descending into submounts
on the client.

In this situation, the `make-snapshot` script that we call from `rsync`
would only create a snapshot of `/backup/host`, not the sub-filesystems that
exist below it.

Instead of changing the script to always create recursive snapshots by
calling `zfs snapshot -r`, I made `make-snapshot` pass all arguments from
its own command line on to `zfs snapshot`; so you can specify the `-r` in
`rsyncd.conf` if you need it.