client/check-if-changed-since-snapshot: re-conceive as a pre-client script
It will now always exit successfully, but create the 'no-recursion' flag file
in the sources.d directory if the source filesystem hasn't changed since the
last backup. `zfsbackup-client`, when it sees this file, doesn't pass
`--recursive` to `rsync` so that `rsync` only actually copies the metadata of
`./path/.`, without traversing the filesystem.

This has the obvious advantage of saving a lot of I/O and real time; but
while the earlier approach where `check-if-changed-since-snapshot` was a
`check` script, it will still allow `rsync` to run and a server-side snapshot
to be created. Thus, no backup snapshots will be "missing" when no backup
takes place because there have been no changes.