zfs-set-korn-zfsbackup-config: rely on inherited $zbZFSDATASET instead of heuristics
Previously, `client/mksource.d/zfs-set-korn-zfsbackup-config` did some voodoo
on its own to find out whether a backup of a zfs instance was being set up.
If yes, it'd add a property to point to the pertinent sources.d directory.
This commit simplifies the code by removing the voodoo and relying on the
parent (`zfsbackup-create-source`) setting `$zbZFSDATASET` to the name of
the zfs dataset being backed up, if any.

The corresponding change to `zfsbackup-create-source` is in the next commit.