Add: server/zfsbackup-restore-postexec, server/zfsbackup-restore-preexec
These scripts can be used to expose all server-side snapshots of a
filesystem as subdirectories of an `rsync` restore module.

If the filesystem has children, and the client accesses a snapshot, the
pertinent snapshots of the child filesystems will be bind mounted under the
correct directories of the root snapshot, so that the client sees the entire
zfs subtree in the state it was in when the recursive snapshot was taken.