Add: client/create-and-mount-snapshot, client/umount-and-destroy-snapshot
These scripts can create zfs and lvm snapshots before a backup is taken, and
destroy them afterwards.

The `lvm` support is untested. Recursive `zfs` snapshots are supported but
untested. `zvol` snapshots are not implemented (the code is just a
copy-pasted placeholder).

The `destroy` script always keep the last snapshot and only destroys
previous ones; this way the `check-if-changed-since-snapshot` `pre-client`
script can reliably determine whether a backup needs to be taken or not.