Add `max connections = 1` to generated `rsyncd.conf` stanzas
For write-only rsync modules this is necessary because snapshot consistency
can't be guaranteed if more than one client is writing to the same module
concurrently.

For read-only rsync modules it's necessary when mounting a recursive
structure of snapshots on demand, so that we can unmount the structure when
the client disconnects.