create-and-mount-snapshot: add explicit support for subsources.d.
If the subdir exists, recursive snapshots will be mounted under the path/
subdir of each appropriate subsource, and snapshots of filesystems that
don't have a subsource will be destroyed before the backup begins. It was
also necessary to introduce a zfs-dataset-root control file that defines the
root of the snapshot hierarchy.

Also:

 * Add umount_fs() helper function because the umount code is needed in more than one place.
 * Fix bug in recursive_zfs_snapshot_postclient() that would cause last successful snapshot to also be removed when snapshots were taken recursively.
 * Improve comments.